+++ samples/SampleEffects/State/Effects/LoadDataEffect.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
-using NetRx.Effects;
+using NetRx.Store.Effects;
 using SampleEffects.State.Reducers;
 using actions = SampleEffects.State.Actions;