+++ src/NetRx.Store.Monitor.Extension/Logic/IOutputPaneParser.cs
@@ -1,0 +1,9 @@
+﻿using EnvDTE;
+
+namespace NetRx.Store.Monitor.Logic
+{
+    public interface IOutputPaneParser
+    {
+        string[] GetLastLines(TextDocument textDocument);
+    }
+}