+++ src/NetRx.Store.Monitor.Shared/Constants.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace NetRx.Store.Monitor.Shared
+{
+    public static class Constants
+    {
+        public static class Message
+        {
+            public const string Tag = "netrx.store";
+        }
+    }
+}