+++ src/NetRx.Store.Monitor.Shared/Models/TraceMessage.cs
@@ -1,0 +1,11 @@
+﻿namespace NetRx.Store.Monitor.Shared.Models
+{
+    public class TraceMessage
+    {
+        public string StoreTypeName { get; set; }
+
+        public string ActionTypeName { get; set; }
+
+        public string StateValueJson { get; set; }
+    }
+}