+++ src/NetRx.Store.Tests/Extensions.cs
@@ -1,8 +1,7 @@
-﻿using NetRx.Effects;
-using System;
-using System.Collections;
+﻿using System;
 using System.Collections.Generic;
 using System.Reflection;
+using NetRx.Store.Effects;
 
 namespace NetRx.Store.Tests
 {