+++ src/NetRx.Store/Diagnostic/ITraceMessageWriter.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace NetRx.Store.Diagnostic
+{
+    internal interface ITraceMessageWriter
+    {
+        void Write(string actionTypeName, IList<StoreItem> states);
+    }
+}